проверки на возможность выполнения запроса с точки зрения прав доступа

+++ backend/src/main/java/com/chat/server/controller/RoomController.java
@@ -4,6 +4,8 @@ import com.chat.server.exception.ObjectNotFoundException;
 import com.chat.server.model.Role;
 import com.chat.server.model.Room;
 import com.chat.server.model.User;
+import com.chat.server.oauth2.domain.UserResource;
+import com.chat.server.oauth2.service.AccessService;
 import com.chat.server.service.RoomService;
 import com.chat.server.service.UserService;
 import org.jboss.logging.Logger;
@@ -13,6 +15,8 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import javax.annotation.security.RolesAllowed;
 
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.bind.annotation.*;
 
@@ -31,6 +35,8 @@ public class RoomController {
     private RoomService roomService;
     @Autowired
     private UserService userService;
+    @Autowired
+    private AccessService accessService;
 
     private static final Logger logger = Logger.getLogger(RoomController.class);
 
@@ -38,6 +44,7 @@ public class RoomController {
      * Get rooms
      * @return HttpEntity<List<Room>> - all rooms that are in system
      */
+    @RolesAllowed({Role.ADMIN})
     @RequestMapping(method = RequestMethod.GET)
     public HttpEntity<List<Room>> getRooms(){
         List<Room> rooms = roomService.findAll();
@@ -62,59 +69,87 @@ public class RoomController {
     }
 
     /**
-     * Get rooms by userId
+     * Get rooms by userId. Only room owner can do this request.
      * @param userId
      * @return HttpEntity<List<Room>> - all rooms that user takes part in
      */
+    // todo: проверить, что пользователь запрашивает свои комнаты -- протестировать
     @RequestMapping(value="/byUserId/{userId}", method = RequestMethod.GET)
     public HttpEntity<List<Room>> getRoomsByUserId(@PathVariable("userId") int userId) throws ObjectNotFoundException{
+        UserResource currentUser = accessService.getCurrentUser();
+        if( currentUser == null || currentUser.getId() != userId ){
+            return new ResponseEntity( HttpStatus.FORBIDDEN );
+        }
+
         List<Room> rooms = userService.findRoomsWithOwnersByUserId(userId);
         return new ResponseEntity( rooms , HttpStatus.OK );
     }
 
     /**
-     * Get room by its id
+     * Get room by its id. Only room owner can do this request.
      * @param id
      * @return HttpEntity<Room> - room
      */
-    // todo: проверить, что комнату запрашивает ее владелец
+    // todo: проверить, что комнату запрашивает ее владелец -- тестировать
     @RequestMapping(value = "/{id}", method = RequestMethod.GET)
     public HttpEntity<Room> getRoom(@PathVariable("id") int id){
         Room room = roomService.findOne(id);
         if (room != null){
+            UserResource currentUser = accessService.getCurrentUser();
+            if( currentUser == null || currentUser.getId() != room.getOwner().getId() ){
+                return new ResponseEntity( HttpStatus.FORBIDDEN );
+            }
+
             return new ResponseEntity( room, HttpStatus.OK );
         }
+
         return new ResponseEntity( HttpStatus.BAD_REQUEST );
     }
 
     /**
-     * Get room's users by its id
+     * Get room's users by its id. Only room owner can do this request.
      * @param id
      * @return HttpEntity<List<User>> - list of users
      */
-    // todo: проверить, что пользователей запрашивает владелец комнаты
+    // todo: проверить, что пользователей запрашивает владелец комнаты -- тестировать
     @RequestMapping(value = "/{id}/users", method = RequestMethod.GET)
     public HttpEntity<List<User>> getRoomUsers(@PathVariable("id") int id) throws ObjectNotFoundException{
-        List<User> users = roomService.getRoomUsers( id );
-        return new ResponseEntity( users, HttpStatus.OK );
+        Room room = roomService.findOne( id );
+        if( room != null ){
+            UserResource currentUser = accessService.getCurrentUser();
+            if( currentUser == null || currentUser.getId() != room.getOwner().getId() ){
+                return new ResponseEntity( HttpStatus.FORBIDDEN );
+            }
+
+            List<User> users = roomService.getRoomUsers( id );
+            return new ResponseEntity( users, HttpStatus.OK );
+        }
+        return new ResponseEntity( HttpStatus.BAD_REQUEST );
     }
 
+    //todo: протестировать
     /**
-     * Update room
+     * Update room. Only room owner can do this request.
      * @param room
      * @return HttpEntity<Room> - updated room
      */
     @RequestMapping(value = "/{id}", method = RequestMethod.PUT)
     public HttpEntity<Room> updateRoom(@RequestBody Room room){
-       roomService.update( room );
+        UserResource currentUser = accessService.getCurrentUser();
+        if( currentUser == null || currentUser.getId() != room.getOwner().getId() ){
+            return new ResponseEntity( HttpStatus.FORBIDDEN );
+        }
+
+        roomService.update( room );
         if (room != null){
             return new ResponseEntity( room, HttpStatus.OK );
         }
         return new ResponseEntity( HttpStatus.BAD_REQUEST );
     }
 
+    //todo: протестировать
     /**
-     * Delete room
+     * Delete room. Only room owner can do this request.
      * @param id
      * @return HttpEntity<String> - if all is ok then HttpStatus.NO_CONTENT else HttpStatus.BAD_REQUEST
      */
@@ -122,9 +157,15 @@ public class RoomController {
     public HttpEntity<String> deleteRoom(@PathVariable("id") int id){
         Room room = roomService.findOne( id );
         if ( room != null ){
+            UserResource currentUser = accessService.getCurrentUser();
+            if( currentUser == null || currentUser.getId() != room.getOwner().getId() ){
+                return new ResponseEntity( HttpStatus.FORBIDDEN );
+            }
+
             roomService.delete( room );
             return new ResponseEntity( HttpStatus.NO_CONTENT );
         }
+
         return new ResponseEntity( HttpStatus.BAD_REQUEST );
     }
 
@@ -168,24 +209,31 @@ public class RoomController {
      */
     @RequestMapping(value="/leave/{id}", method = RequestMethod.POST)
     public HttpEntity<Room> leaveRoom(@PathVariable("id") int roomId , @RequestBody int userId) throws ObjectNotFoundException{
-        User user = userService.findOne( userId );
-        Room room = roomService.removeUserFromRoom( roomId, user );
+        User user = userService.findOne(userId);
+        roomService.removeUserFromRoom( roomId, user );
         return new ResponseEntity( HttpStatus.NO_CONTENT );
     }
 
+    //todo: протестировать
     /**
-     * Remove a list of users from room's users
+     * Remove a list of users from room's users. Only room owner can do this request.
      * @param roomId
      * @param usersToRemove
      * @return HttpEntity<Room> - HttpStatus.NO_CONTENT when all is  ok
      */
     @RequestMapping(value="/removeUsers/{id}", method = RequestMethod.POST)
     public HttpEntity<Room> removeUsers(@PathVariable("id") int roomId, @RequestBody List<User> usersToRemove) throws ObjectNotFoundException{
-        Room room = roomService.removeUsersFromRoom( roomId, usersToRemove );
-        if ( room == null ){
-            return new ResponseEntity( HttpStatus.BAD_REQUEST );
+        Room room = roomService.findOne(roomId);
+        if( room != null ){
+            UserResource currentUser = accessService.getCurrentUser();
+            if( currentUser == null || currentUser.getId() != room.getOwner().getId() ){
+                return new ResponseEntity( HttpStatus.FORBIDDEN );
+            }
+            roomService.removeUsersFromRoom(roomId, usersToRemove);
+            return new ResponseEntity( HttpStatus.NO_CONTENT );
         }
-        return new ResponseEntity( HttpStatus.NO_CONTENT );
+
+        return new ResponseEntity( HttpStatus.BAD_REQUEST );
     }
 
 }
